+++ csvviz/cmds/scatter.py
@@ -52,11 +52,8 @@ class ScatterKit(Vizkit):
     def __init__(self, input_file, kwargs):
         super().__init__(viz_type="scatter", input_file=input_file, kwargs=kwargs)
 
-    def prepare_channels(
-        self,
-    ):  # -> typeDict[str, typeUnion[alt.X, alt.Y, alt.Fill, alt.Size]]:
-
-        channels = self._init_channels(self.channel_kwargs, self.datakit)
+    def prepare_channels(self):
+        channels = self._channels_init(self.channel_kwargs, self.datakit)
         if channels.get("fill"):
             channels["fill"].scale = alt.Scale(**self._config_colors(self.color_kwargs))